[Core] uint24_t and int24_t as full supported types
