+++ Source/core/core.h
@@ -109,6 +109,7 @@
 #include "WarningReportingCategories.h"
 #include "Number.h"
 #include "ExtraNumberDefinitions.h"
+#include "Errors.h"
 
 #ifdef __WINDOWS__
 #pragma comment(lib, "core.lib")
